# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        example: Error message here
        type: string
    type: object
...
  models.Invoice:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      order_id:
        example: 507f1f77bcf86cd799439012
        type: string
      payment_method:
        enum:
        - cash
        - credit_card
        - debit_card
        - mobile_payment
        example: credit_card
        type: string
      payment_status:
        enum:
        - pending
        - paid
        - failed
        - refunded
        example: paid
        type: string
      total_amount:
        example: 45.99
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - order_id
    - payment_method
    - payment_status
    - total_amount
    type: object
  models.InvoiceCreateRequest:
    properties:
      order_id:
        example: 507f1f77bcf86cd799439012
        type: string
      payment_method:
        enum:
        - cash
        - credit_card
        - debit_card
        - mobile_payment
        example: credit_card
        type: string
      payment_status:
        enum:
        - pending
        - paid
        - failed
        - refunded
        example: paid
        type: string
      total_amount:
        example: 450
        type: number
    required:
    - order_id
    - payment_method
    - payment_status
    - total_amount
    type: object
  models.InvoiceResponse:
    properties:
      id:
        example: 507f1f77bcf86cd799439018
        type: string
      invoice:
        $ref: '#/definitions/models.Invoice'
      message:
        example: Invoice fetched successfully
        type: string
    type: object
...
  models.Menu:
    properties:
      category:
        example: Main Course
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        example: Dinner Menu
        type: string
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - category
    - name
    type: object
  models.MenuCreateRequest:
    properties:
      category:
        example: Main Course
        type: string
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      name:
        example: Dinner Menu
        type: string
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - category
    - name
    type: object
  models.MenuResponse:
    properties:
      id:
        example: 507f1f77bcf86cd799439014
        type: string
      menu:
        $ref: '#/definitions/models.Menu'
      message:
        example: Menu fetched successfully
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant.com
    name: API Support
  description: This is a restaurant management system API with user authentication,
    food ordering, and invoice management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
paths:
...
  /invoices:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all invoices
      produces:
      - application/json
      responses:
        "200":
          description: List of invoices
          schema:
            items:
              $ref: '#/definitions/models.InvoiceResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Invoices
      tags:
      - Invoice
    post:
      consumes:
      - application/json
      description: Create a new invoice
      parameters:
      - description: Invoice details
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invoice created successfully
          schema:
            $ref: '#/definitions/models.InvoiceResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Invoice
      tags:
      - Invoice
  /invoices/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific invoice by its ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice details
          schema:
            $ref: '#/definitions/models.InvoiceResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Invoice by ID
      tags:
      - Invoice
    put:
      consumes:
      - application/json
      description: Update an existing invoice
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated invoice details
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invoice updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Invoice not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Invoice
      tags:
      - Invoice
  /menus:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all menus
      produces:
      - application/json
      responses:
        "200":
          description: List of menus
          schema:
            items:
              $ref: '#/definitions/models.MenuResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Menus
      tags:
      - Menu
    post:
      consumes:
      - application/json
      description: Create a new menu
      parameters:
      - description: Menu details
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Menu created successfully
          schema:
            $ref: '#/definitions/models.MenuResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Menu
      tags:
      - Menu
  /menus/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Menu deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Menu
      tags:
      - Menu
    get:
      consumes:
      - application/json
      description: Retrieve a specific menu by its ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Menu details
          schema:
            $ref: '#/definitions/models.MenuResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Menu by ID
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: Update an existing menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated menu details
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Menu updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"